+++ src/main/java/com/swpproject/application/repository/PersonalTrainerRepository.java
@@ -3,5 +3,5 @@ package com.swpproject.application.repository;
 import com.swpproject.application.model.PersonalTrainer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonalTrainerRepository extends JpaRepository<PersonalTrainer, Long> {
+public interface PersonalTrainerRepository extends JpaRepository<PersonalTrainer, Integer> {
 }
